+++ gargoyle/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+
+class GargoyleAppConfig(AppConfig):
+    name = 'gargoyle'
+    verbose_name = 'Gargoyle'
+
+    def ready(self):
+        self.module.autodiscover()