+++ HISTORY.rst
@@ -8,6 +8,9 @@ Pending Release
 ---------------
 
 * New release notes here
+* Simplified autodiscovery code to use ``AppConfig.ready()``. It's no longer
+  necessary to add a call to ``gargoyle.autodiscover()`` in your ``urls.py``,
+  when not using Nexus.
 
 1.2.0 (2016-02-12)
 ------------------